Allow query parameter check on mocked url Split Classes in Router in diferent files
