+++ Sources/Vaporiser/Router.swift
@@ -79,90 +79,3 @@ class Router {
         return mock
     }
 }
-
-public struct MockData: Codable {
-
-    public let path: String
-    public let payload: Data?
-    public let method: Method
-    public let returnCode: Int
-
-    public init(path: String, payload: Data? = nil, method: MockData.Method, returnCode: Int = 200) {
-        self.path = path
-        self.payload = payload
-        self.method = method
-        self.returnCode = returnCode
-    }
-
-    var pathComponents: [PathComponent] {
-        path.pathComponents
-    }
-
-    func matches(url: URI) -> Bool {
-        path.matches(url: url)
-    }
-
-    func matches(url: URI, method: Method) -> Bool {
-        guard method == self.method else { return false }
-        return matches(url: url)
-    }
-
-    public enum Method: Codable, Equatable {
-        case GET
-        case PUT
-        case HEAD
-        case POST
-        case PATCH
-        case DELETE
-        case OTHER(String)
-
-        init(httpMethod: HTTPMethod) {
-            switch httpMethod {
-            case .GET: self = .GET
-            case .PUT: self = .PUT
-            case .HEAD: self = .HEAD
-            case .POST: self = .POST
-            case .DELETE: self = .DELETE
-            default: self = .OTHER(httpMethod.rawValue)
-            }
-        }
-    }
-}
-
-private extension String {
-    func matches(url: URI) -> Bool {
-        guard url.path.pathComponents.count == pathComponents.count else {
-            return false
-        }
-
-        var match = true
-        for (component1, component2) in zip(url.path.pathComponents, pathComponents) {
-
-            if component2 == .anything {
-                continue
-            }
-
-            if component1 != component2 {
-                match = false
-                break
-            }
-        }
-
-        return match
-    }
-}
-
-private extension Dictionary where Key == String, Value == MockData {
-    func firstMatch(url: URI) -> MockData? {
-        guard let universalPath = keys.first(where: { $0.matches(url: url)})  else {
-            return nil
-        }
-        return self[universalPath]
-    }
-}
-
-public extension Array where Element == MockData {
-    func firstMatch(url: URI, method: HTTPMethod) -> MockData? {
-        first(where: { $0.matches(url: url, method: .init(httpMethod: method))})
-    }
-}
